Roulette/requirements.txt
pytest
Appium-Python-Client
AltTester-Driver==2.0.2
